requirements.txt
openai_whisper==20230314
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pydub==0.25.1
python-dotenv==1.0.1
